+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - May 21, 2025: Site is live!
 - May 23, 2025: Modify Dijkstra to use `long long`, add `MOD` to Fast Exponentiation
 - May 26, 2025: Add alternative to bitset
+- June 9, 2025: Add `const int N` to template files.
+- June 21, 2025: Add Range Tree on Trees.